refactor: extract tx broadcast from component
